refactor(ivy): combine styling testing files into one
